Refactored validation method to be more succinct. Moved related unit tests to a more appropriate location.
